+++ bin/miden-cli/src/commands/new_account.rs
@@ -391,7 +391,7 @@ async fn create_client_account<AUTH: TransactionAuthenticator + Sync + 'static>(
         None
     } else {
         debug!("Adding default Falcon auth component");
-        let kp = AuthSecretKey::new_rpo_falcon512_with_rng(client.rng());
+        let kp = AuthSecretKey::new_falcon512_rpo_with_rng(client.rng());
         builder =
             builder.with_auth_component(AuthRpoFalcon512::new(kp.public_key().to_commitment()));
         Some(kp)