+++ crates/web-client/src/models/code_builder.rs
@@ -1,20 +1,21 @@
 use alloc::sync::Arc;
 
-use miden_client::CodeBuilder as NativeCodeBuilder;
+use miden_client::account::AccountComponentCode as NativeAccountComponentCode;
 use miden_client::assembly::{
     Assembler,
+    CodeBuilder as NativeCodeBuilder,
     Library as NativeLibrary,
-    LibraryPath,
     Module,
     ModuleKind,
+    Path,
     PrintDiagnostic,
     Report,
     SourceManagerSync,
 };
-use miden_client::transaction::TransactionKernel;
 use wasm_bindgen::prelude::*;
 
 use crate::js_error_with_context;
+use crate::models::account_component_code::AccountComponentCode;
 use crate::models::library::Library;
 use crate::models::note_script::NoteScript;
 use crate::models::transaction_script::TransactionScript;
@@ -28,17 +29,13 @@ pub struct CodeBuilder {
     // need an assembler to compile an AccountComponent. After miden-base issue #1756 is complete
     // we will be able to remove the Assembler and only use the NativeCodeBuilder.
     builder: NativeCodeBuilder,
-    assembler: Assembler,
 }
 
 #[wasm_bindgen]
 impl CodeBuilder {
     pub(crate) fn from_source_manager(source_manager: Arc<dyn SourceManagerSync>) -> Self {
         let builder = NativeCodeBuilder::with_source_manager(source_manager);
-        let assembler = TransactionKernel::assembler_with_source_manager(builder.source_manager().clone())
-                // When instanced with a source manager, the builder has debug mode on by default.
-                .with_debug_mode(true);
-        Self { builder, assembler }
+        Self { builder }
     }
 
     /// Given a module path (something like `my_lib::module`) and source code, this will
@@ -50,11 +47,6 @@ impl CodeBuilder {
                 e,
                 &format!("script builder: failed to link module with path {module_path}"),
             )
-        })?;
-        self.assembler.compile_and_statically_link(module_code).map_err(|e| {
-            let err_msg =
-                format_assembler_error(&e, "script builder: assembler failed to link module");
-            JsValue::from(err_msg)
         })?;
         Ok(())
     }
@@ -71,13 +63,6 @@ impl CodeBuilder {
         self.builder.link_static_library(&library).map_err(|e| {
             js_error_with_context(e, "script builder: failed to link static library")
         })?;
-        self.assembler.link_static_library(&library).map_err(|e| {
-            let err_msg = format_assembler_error(
-                &e,
-                "script builder: assembler failed to link static library",
-            );
-            JsValue::from_str(&err_msg)
-        })?;
         Ok(())
     }
 
@@ -92,13 +77,6 @@ impl CodeBuilder {
         let library: NativeLibrary = library.into();
         self.builder.link_dynamic_library(&library).map_err(|e| {
             js_error_with_context(e, "script builder: failed to link dynamic library")
-        })?;
-        self.assembler.link_dynamic_library(&library).map_err(|e| {
-            let err_msg = format_assembler_error(
-                &e,
-                "script builder: assembler failed to link dynamic library",
-            );
-            JsValue::from_str(&err_msg)
         })?;
         Ok(())
     }
@@ -133,22 +111,24 @@ impl CodeBuilder {
     /// Given a Library Path, and a source code, turn it into a Library.
     /// E.g. A path library can be `miden::my_contract`. When turned into a library,
     /// this can be used from another script with an import statement, following the
-    /// previous example: `use.miden::my_contract'.
+    /// previous example: `use miden::my_contract'.
     #[wasm_bindgen(js_name = "buildLibrary")]
     pub fn build_library(&self, library_path: &str, source_code: &str) -> Result<Library, JsValue> {
-        let library_path = LibraryPath::new(library_path).map_err(|e| {
+        let library_path = Path::validate(library_path).map_err(|e| {
             js_error_with_context(
-                e, "script builder: failed to build library -- could not create library_path with path {library_path}",
+                e,
+                &format!("script builder: failed to build library -- invalid path {library_path}"),
             )
         })?;
         let module = Module::parser(ModuleKind::Library)
-            .parse_str(library_path, source_code, self.builder.source_manager().as_ref())
+            .parse_str(library_path, source_code, self.builder.source_manager())
             .map_err(|e| {
                 let err_msg = format_assembler_error(&e, "error while parsing module");
                 JsValue::from(err_msg)
             })?;
 
-        let native_library_build = self.clone_assembler().assemble_library([module]);
+        let assembler: Assembler = self.builder.clone().into();
+        let native_library_build = assembler.assemble_library([module]);
         match native_library_build {
             Ok(native_library) => Ok(native_library.into()),
             Err(error_report) => {
@@ -159,10 +139,23 @@ impl CodeBuilder {
         }
     }
 
-    /// Returns the inner assembler . This is because multiple "compile" functions
-    /// in `miden_lib` to consume an Assembler, so we need to clone the value.
-    pub(crate) fn clone_assembler(&self) -> Assembler {
-        self.assembler.clone()
+    /// Given an `AccountComponentCode`, compiles it
+    /// with the available modules under this builder. Returns the compiled account component code.
+    #[wasm_bindgen(js_name = "compileAccountComponentCode")]
+    pub fn compile_account_component_code(
+        &self,
+        account_code: &str,
+    ) -> Result<AccountComponentCode, JsValue> {
+        let assembler: Assembler = self.builder.clone().into();
+        let native_library = assembler.assemble_library([account_code]).map_err(|e| {
+            JsValue::from_str(&format!(
+                "Failed to compile account component:
+        {e}"
+            ))
+        })?;
+
+        let native_code: NativeAccountComponentCode = native_library.into();
+        Ok(native_code.into())
     }
 }