+++ crates/web-client/src/new_transactions.rs
@@ -9,8 +9,8 @@ use miden_client::transaction::{
 };
 use wasm_bindgen::prelude::*;
 
+use crate::models::NoteType;
 use crate::models::account_id::AccountId;
-use crate::models::note_type::NoteType;
 use crate::models::proven_transaction::ProvenTransaction;
 use crate::models::provers::TransactionProver;
 use crate::models::transaction_id::TransactionId;
@@ -23,6 +23,8 @@ use crate::{WebClient, js_error_with_context};
 impl WebClient {
     /// Executes a transaction specified by the request against the specified account,
     /// proves it, submits it to the network, and updates the local database.
+    ///
+    /// Uses the prover configured for this client.
     ///
     /// If the transaction utilizes foreign account data, there is a chance that the client doesn't
     /// have the required block header in the local database. In these scenarios, a sync to
@@ -38,6 +40,33 @@ impl WebClient {
         let tx_id = transaction_result.id();
 
         let proven_transaction = self.prove_transaction(&transaction_result, None).await?;
+
+        let submission_height =
+            self.submit_proven_transaction(&proven_transaction, &transaction_result).await?;
+        self.apply_transaction(&transaction_result, submission_height).await?;
+
+        Ok(tx_id)
+    }
+
+    /// Executes a transaction specified by the request against the specified account, proves it
+    /// with the user provided prover, submits it to the network, and updates the local database.
+    ///
+    /// If the transaction utilizes foreign account data, there is a chance that the client doesn't
+    /// have the required block header in the local database. In these scenarios, a sync to the
+    /// chain tip is performed, and the required block header is retrieved.
+    #[wasm_bindgen(js_name = "submitNewTransactionWithProver")]
+    pub async fn submit_new_transaction_with_prover(
+        &mut self,
+        account_id: &AccountId,
+        transaction_request: &TransactionRequest,
+        prover: &TransactionProver,
+    ) -> Result<TransactionId, JsValue> {
+        let transaction_result = self.execute_transaction(account_id, transaction_request).await?;
+
+        let tx_id = transaction_result.id();
+
+        let proven_transaction =
+            self.prove_transaction(&transaction_result, Some(prover.clone())).await?;
 
         let submission_height =
             self.submit_proven_transaction(&proven_transaction, &transaction_result).await?;